fixing hf 15 * add hf list 15 * refactoring: move all of hf list code to cmdhflist.c
