+++ client/cmdsmartcard.c
@@ -17,7 +17,7 @@
 #include "smartcard.h"
 #include "comms.h"
 #include "protocols.h"
-#include "cmdhf.h"              // CmdHFlist
+#include "cmdhflist.h"
 #include "emv/apduinfo.h"       // APDUcode description
 #include "emv/emvcore.h"        // decodeTVL
